requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
firebase-admin==6.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.5.0
pyrebase4==4.6.0
python-jose==3.3.0
gunicorn==21.2.0


flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.5.0
python-dotenv==1.0.0
stripe==7.0.0
python-multipart==0.0.6
requests==2.31.0
pyjwt==2.8.0
python-dateutil==2.8.2


# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Firebase Integration
firebase-admin==6.2.0
google-cloud-storage==2.8.0
python-jose==3.3.0
pyrebase4==4.5.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Async & Performance
httpx==0.25.2
aiofiles==23.2.1
asgiref==3.7.2

# Data Processing & Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
uuid==1.30
python-magic==0.4.27

# Excel/PDF Export (For Reports)
openpyxl==3.1.2
xlsxwriter==3.1.2
reportlab==4.0.6
pandas==2.1.3

# Email & Notifications
smtplib==1.0.1
python-dotenv==1.0.0
requests==2.31.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
black==23.11.0
flake8==6.1.0

# Environment Management
python-dotenv==1.0.0